Tidy up graph tests (#69)
* move fixtures out of test file

* split test_graph.py into three files

* factory as fixture (see https://docs.pytest.org/en/7.1.x/how-to/fixtures.html#factories-as-fixtures)

* more fixtures

* typing

* Remove model_constructor and replace with just model that takes parameter args

* Add option to change run args in MCMC fixtures

* Update the tests to use the simplier framework

* Rename test file in line with new name

* Fix missing default case

* Apply suggestions from code review


* Ruff giveth, ruff taketh away

* Remove doctest that is actually now just a test

* update tests and fixtures

* ruffff

* update tests and sampling code

* Apply suggestions from code review


* ruff

* mypy

* mypy?

---------
