+++ tests/test_causal_problem/conftest.py
@@ -1,6 +1,6 @@
 import pytest
+from numpyro.distributions import Normal
 
-from causalprog.distribution.normal import NormalFamily
 from causalprog.graph import DistributionNode, Graph, ParameterNode
 
 
@@ -18,16 +18,16 @@ def graph() -> Graph:
 
     mu_x = ParameterNode(label="mu_x")
     x = DistributionNode(
-        NormalFamily(),
+        Normal,
         label="x",
-        parameters={"mean": "mu_x"},
-        constant_parameters={"cov": 1.0},
+        parameters={"loc": "mu_x"},
+        constant_parameters={"scale": 1.0},
     )
     nu_y = ParameterNode(label="nu_y")
     y = DistributionNode(
-        NormalFamily(),
+        Normal,
         label="y",
-        parameters={"mean": "x", "cov": "nu_y"},
+        parameters={"loc": "x", "scale": "nu_y"},
     )
 
     graph.add_edge(mu_x, x)
