+++ tests/test_graph/test_model.py
@@ -1,64 +1,21 @@
-from collections.abc import Callable
-
 import numpy.typing as npt
 import numpyro
 import pytest
-from numpyro.distributions import Normal
 
 from causalprog.graph import DistributionNode, Graph, ParameterNode
-
-
-@pytest.fixture
-def two_normal_graph() -> Graph:
-    g = Graph(label="Two Normals")
-
-    mu_x = ParameterNode(label="mu_x")
-    nu_y = ParameterNode(label="nu_y")
-
-    x = DistributionNode(
-        Normal,
-        label="X",
-        parameters={"loc": "mu_x"},
-        constant_parameters={"scale": 1.0},
-    )
-    y = DistributionNode(
-        Normal,
-        label="Y",
-        parameters={"loc": "X", "scale": "nu_y"},
-    )
-    g.add_node(mu_x)
-    g.add_node(nu_y)
-    g.add_edge(mu_x, x)
-    g.add_edge(nu_y, y)
-    g.add_edge(x, y)
-
-    return g
-
-
-@pytest.fixture
-def two_normal_graph_expected_model() -> Callable[..., dict[str, npt.ArrayLike]]:
-    """Creates the model that the two_normal_graph should produce."""
-
-    def _inner(mu_x: float, nu_y: float) -> dict[str, npt.ArrayLike]:
-        x = numpyro.sample("X", Normal(loc=mu_x, scale=1.0))
-        y = numpyro.sample("Y", Normal(loc=x, scale=nu_y))
-
-        return {"X": x, "Y": y}
-
-    return _inner
 
 
 @pytest.mark.parametrize(
     "param_values",
     [
-        pytest.param({"mu_x": 0.0, "nu_y": 1.0}, id="mu_x = 0, nu_y = 1"),
-        pytest.param({"mu_x": 1.0, "nu_y": 2.0}, id="mu_x = 1, nu_y = 2"),
+        pytest.param({"mu_x": 0.0, "nu_y": 1.0}, id="mean(X) = 0, cov(UX) = 1"),
+        pytest.param({"mu_x": 1.0, "nu_y": 2.0}, id="mean(X) = 1, cov(UX) = 2"),
     ],
 )
 def test_model(
     param_values: dict[str, npt.ArrayLike],
-    two_normal_graph: Graph,
-    two_normal_graph_expected_model: Callable[..., dict[str, npt.ArrayLike]],
+    two_normal_graph_parametrized_mean,
+    two_normal_graph_expected_model,
     assert_samples_are_identical,
     run_default_nuts_mcmc,
 ) -> None:
@@ -74,10 +31,11 @@ def test_model(
     MCMC sampling output of `Graph.model` with the explicit model that it should
     correspond to.
     """
-    assert callable(two_normal_graph.model)
+    graph = two_normal_graph_parametrized_mean()
+    assert callable(graph.model)
 
     via_model = run_default_nuts_mcmc(
-        two_normal_graph.model,
+        graph.model,
         mcmc_run_kwargs=param_values,
     )
     via_expected = run_default_nuts_mcmc(
@@ -89,13 +47,15 @@ def test_model(
 
 
 def test_model_missing_parameter(
-    two_normal_graph: Graph,
+    two_normal_graph_parametrized_mean,
     raises_context,
     seed: int,
 ) -> None:
     """`Graph.model` will raise a `KeyError` when a value is not passed for
     a `ParameterNode`.
     """
+    graph = two_normal_graph_parametrized_mean()
+
     # Deliberately leave out the "nu_y" variable.
     parameter_values = {"mu_x": 0.0}
     # Which should result in the error below.
@@ -103,22 +63,24 @@ def test_model_missing_parameter(
 
     # Not passing enough parameters should be picked up by the model.
     with raises_context(expected_exception), numpyro.handlers.seed(rng_seed=seed):
-        two_normal_graph.model(**parameter_values)
+        graph.model(**parameter_values)
 
 
 def test_model_extension(
-    two_normal_graph: Graph,
+    two_normal_graph_parametrized_mean,
     assert_samples_are_identical,
     run_default_nuts_mcmc,
 ) -> None:
     """Test that `Graph.model` can be extended."""
+    graph = two_normal_graph_parametrized_mean()
+
     parameter_values = {"mu_x": 0.0, "nu_y": 1.0}
 
-    # Build the two_normal_graph, but without the Y-node.
+    # Build the graph, but without the X-node.
     mu_x = ParameterNode(label="mu_x")
     x = DistributionNode(
         numpyro.distributions.Normal,
-        label="X",
+        label="UX",
         parameters={"loc": "mu_x"},
         constant_parameters={"scale": 1.0},
     )
@@ -128,15 +90,15 @@ def test_model_extension(
     def extended_model(*, nu_y, **parameter_values):
         sites = one_normal_graph.model(**parameter_values)
         numpyro.sample(
-            "Y",
+            "X",
             numpyro.distributions.Normal(
-                loc=sites["X"],
+                loc=sites["UX"],
                 scale=nu_y,
             ),
         )
 
     via_explicit_graph = run_default_nuts_mcmc(
-        two_normal_graph.model, mcmc_run_kwargs=parameter_values
+        graph.model, mcmc_run_kwargs=parameter_values
     )
     via_extended = run_default_nuts_mcmc(
         extended_model, mcmc_run_kwargs=parameter_values
